% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirna_base_frequency.R
\name{basepreplot}
\alias{basepreplot}
\title{Generate the base frequency plot of miRNA}
\usage{
basepreplot(miRNAdata, width = 0.6, text_size = 10, title_size = 12)
}
\arguments{
\item{miRNAdata}{A dataframe. The output result after running mirnapredata.}

\item{width}{A numeric. The width of the output bar plot, and default is 0.6.}

\item{text_size}{A numeric. The size of text in the output plot, and default is 10.}

\item{title_size}{A numeric. The size of title text in the output plot, and default is 12.}
}
\value{
The miRNA base frequency plot.
}
\description{
Generate the base frequency plot of miRNA
}
\examples{
#F1
F1_miRNA <- F1_miRNA_count[,1]
F1_bf <- mirnapredata(miRNAseq = F1_miRNA)
basepreplot(miRNAdata = F1_bf)
}
