% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{F1_miRNA_count}
\alias{F1_miRNA_count}
\title{Count table of miRNAs in F1 (F1: the polyploid progeny).}
\description{
Count table of miRNAs in F1 species. The "F1" represents the polyploid progeny.
}
\examples{
head(F1_miRNA_count)
#                sequence     BF1.1    BF1.2    BF1.3
#1  TTTGGATTGAAGGGAGCTCTA     20233     6388    16732
#2  TTTCCAAATGTAGACAAAGCA     19909     5157    16076
#3    TCCCAAATGTAGACAAAGC        82       33      103
#4 CTTTGTCTATCGTTTGGAAAAG      2367     1040     3203
#5  TTGGACTGAAGGGAGCTCCTT        34        9       21
#6  TCGGACCAGGCTTCATTCCCC      3281      607     1289
}
