% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{F1_miRNA_rpm}
\alias{F1_miRNA_rpm}
\title{RPM table of miRNAs in F1 (F1: the polyploid progeny).}
\description{
RPM table of miRNAs in F1 species. The "F1" represents the polyploid progeny.
}
\examples{
head(F1_miRNA_rpm)
#                sequence     BF1.1    BF1.2     BF1.3
#1  TTTGGATTGAAGGGAGCTCTA   1512.16  1086.35   2032.97
#2  TTTCCAAATGTAGACAAAGCA   1487.94   877.01   1953.27
#3    TCCCAAATGTAGACAAAGC      6.13     5.61     12.51
#4 CTTTGTCTATCGTTTGGAAAAG    176.90   176.86    389.17
#5  TTGGACTGAAGGGAGCTCCTT      2.54     1.53      2.55
#6  TCGGACCAGGCTTCATTCCCC    245.21   103.23    156.62
}
