% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get12Bins.R
\name{Get12Bins}
\alias{Get12Bins}
\title{Genetic effects analysis: Twelve bins of expression analysis (method2)}
\usage{
Get12Bins(
  P1_count,
  P2_count,
  F1_count,
  type,
  homoeologs,
  count_threshold = 5,
  Pvalue = 0.05,
  lg2FC = 1
)
}
\arguments{
\item{P1_count}{A dataframe. The count data of miRNA from the P1 species. The first column must be the miRNA sequence. Others are listed as the count of miRNA, and each column denotes one biological replicate of the sample.}

\item{P2_count}{A dataframe. Similar with P1_count, the count data of miRNA from the P2 species.}

\item{F1_count}{A dataframe. Similar with P1_count, the count data of miRNA from the F1 species.}

\item{type}{A character. "sRNA" or "mRNA".}

\item{homoeologs}{A dataframe. if the "type" is "mRNA", the option is required. The first column is the group name (unique) of homoeologs among three species, the second column is the Gene ID of P1, the third column is the Gene ID of P2. And the fourth column is the ID in the F1 generation that is consistent with the P1 genome type, while the fifth column is the ID in the F1 generation that is consistent with the P2 genome type. Such as "Homoeolog1    BraA01t00004Z    BolC01g000040.2J BnA01g0000030.1 BnC01g0424620.1".}

\item{count_threshold}{A numeric. Among the three species(P1,P2,F1), all replicates in at least one species with count values greater than or equal to 5 were retained. By default, the count value more than or equal to 5 is retained.}

\item{Pvalue}{A numeric. The threshold of significance test among different groups. Default is 0.05.}

\item{lg2FC}{A numeric. The fold Change of expression level, and the number is positive. Default is 1.}
}
\value{
A dataframe. The output results contain the P value, log2FoldChange and grouping information for each miRNA/siRNA/gene expressed in all species (count >= count_threshold). F1_vs_P1(P value: pv11,log2FoldChange: fc11), F1_vs_P2(P value: pv12,log2FoldChange: fc12), P2_vs_P1(P value: pv21,log2FoldChange: fc21)
}
\description{
Genetic effects analysis: Twelve bins of expression analysis (method2)
}
\examples{
\dontrun{
miRNA_12bin <- Get12Bins(P1_count = P1_miRNA_count,
                         P2_count = P2_miRNA_count,
                         F1_count = F1_miRNA_count,type = "sRNA")
}
}
