% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDEdata.R
\name{GetDEdata}
\alias{GetDEdata}
\title{GetDEdata: Get specific results of differential expressed genes.}
\usage{
GetDEdata(
  P1_count,
  P2_count,
  F1_count,
  P1_name,
  P2_name,
  F1_name,
  output_type,
  type,
  homoeologs,
  count_threshold = 5,
  Pvalue = 0.05
)
}
\arguments{
\item{P1_count}{A dataframe. The count data of miRNA from the P1 species. The first column must be the miRNA sequence. Others are listed as the count of miRNA, and each column denotes one biological replicate of the sample.}

\item{P2_count}{A dataframe. Similar with P1_count, the count data of miRNA from the P2 species.}

\item{F1_count}{A dataframe. Similar with P1_count, the count data of miRNA from the F1 species.}

\item{P1_name}{A character. The text in the figure represents the name of one of the parents (P1).}

\item{P2_name}{A character. The text in the figure represents the name of one of the parents (P2).}

\item{F1_name}{A character. The text in the figure represents the name of hybrid offspring (F1).}

\item{output_type}{A character. 'F1_vs_P1', 'F1_vs_P2' or 'P2_vs_P1'.}

\item{type}{A character. "sRNA" or "mRNA".}

\item{homoeologs}{A dataframe. if the "type" is "mRNA", the option is required. The first column is the group name (unique) of homoeologs among three species, the second column is the Gene ID of P1, the third column is the Gene ID of P2. And the fourth column is the ID in the F1 generation that is consistent with the P1 genome type, while the fifth column is the ID in the F1 generation that is consistent with the P2 genome type. Such as "Homoeolog1    BraA01t00004Z    BolC01g000040.2J BnA01g0000030.1 BnC01g0424620.1".}

\item{count_threshold}{A numeric. Among the three species(P1,P2,F1), all replicates in at least one species with count values greater than or equal to 5 were retained. By default, the count value more than or equal to 5 is retained.}

\item{Pvalue}{A numeric. The threshold of significance test among different groups. Default is 0.05.}
}
\value{
A dataframe. Differential expression analysis results of miRNA/siRNA/gene expressed in each two species (count >= count_threshold).
}
\description{
GetDEdata: Get specific results of differential expressed genes.
}
\examples{
\dontrun{
P2_vs_P1 <- GetDEdata(P1_count = P1_miRNA_count,
                      P2_count = P2_miRNA_count,
                      F1_count = F1_miRNA_count,
                      P1_name = "B.napus(AACC)",
                      P2_name = "B.rapa(AA)",
                      F1_name = "B.napus x B.rapa (AAAACC)",
                      output_type = "P2_vs_P1", type="sRNA")
}
}
