% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sRNA_length_distribution.R
\name{lenplot}
\alias{lenplot}
\title{Generate the sRNA length distribution plot}
\usage{
lenplot(sRNAdata, type, width = 0.6, text_size = 10, title_size = 12)
}
\arguments{
\item{sRNAdata}{A dataframe. The output result after running "srnapredata".}

\item{type}{A character. "bar" or "line", and specific option will generate specific image.}

\item{width}{A numeric. The width of the output bar plot, and default is 0.6.}

\item{text_size}{A numeric. The size of text in the output plot, and default is 10.}

\item{title_size}{A numeric. The size of title text in the output plot, and default is 12.}
}
\value{
The sRNA length distribution plot.
}
\description{
Generate the sRNA length distribution plot
}
\examples{
#P1(B.napus)
B.napu_sRNA <- srnapredata(sRNAseq = P1_sRNA_seq,Group = "B.napus(AACC)")
#P2(B.rapa)
B.rapa_sRNA <- srnapredata(sRNAseq = P2_sRNA_seq,Group = "B.rapa(AA)")
#F1(B.napus X B.rapa)
B.nr_sRNA <- srnapredata(sRNAseq = F1_sRNA_seq,Group = "B.napus x B.rapa(AAAACC)")
#intergrate these data for length distribution plot
sRNA_data <- rbind(B.napu_sRNA,B.rapa_sRNA,B.nr_sRNA)
#plot
lenplot(sRNAdata = sRNA_data,type = "line")
lenplot(sRNAdata = sRNA_data,type = "bar")
}
