% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P1_miRNA_rpm}
\alias{P1_miRNA_rpm}
\title{RPM table of miRNAs in P1 (P1: one of the parents).}
\description{
RPM table of miRNAs in P1 species. The "P1" represents one of parents.
}
\examples{
head(P1_miRNA_rpm)
#               sequence    Brapa.1    Brapa.2    Brapa.3
#1 TTTGGATTGAAGGGAGCTCTA    1641.18    1116.03    1014.37
#2 TGAAGCTGCCAGCATGATCTA     129.33     103.23     103.68
#3 TTTCCAAATGTAGACAAAGCA     905.23     920.57    1180.51
#4 TCGGACCAGGCTTCATCCCCC      24.71      14.38      15.03
#5 AGAATCTTGATGATGCTGCAG      48.64      41.09      41.60
#6 TTGACAGAAGAAAGAGAGCAC      86.96      81.23      67.41
}
