% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P2_miRNA_count}
\alias{P2_miRNA_count}
\title{Count table of miRNAs in P2 (P2: one of the parents).}
\description{
Count table of miRNAs in P2 species. The "P2" represents one of parents.
}
\examples{
head(P2_miRNA_count)
#                sequence  Bnapus.1  Bnapus.2  Bnapus.3
#1  TTTGGATTGAAGGGAGCTCTA     29848     12094     10685
#2  TTAGATTCACGCACAAACTCG       986       571       456
#3  TGAAGCTGCCAGCATGATCTA      3152      1436      1091
#4 CTTTGTCTATCGTTTGGAAAAG      2449      1307      1116
#5  GATCATGTTCGCAGTTTCACC      1364       650       656
#6  TTTCCAAATGTAGACAAAGCA     11658      3914      4123
}
