% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P2_miRNA_rpm}
\alias{P2_miRNA_rpm}
\title{RPM table of miRNAs in P2 (P2: one of the parents).}
\description{
RPM table of miRNAs in P2 species. The "P2" represents one of parents.
}
\examples{
head(P2_miRNA_rpm)
#                sequence   Bnapus.1   Bnapus.2   Bnapus.3
#1  TTTGGATTGAAGGGAGCTCTA    1804.35    1362.88    1439.22
#2  TTAGATTCACGCACAAACTCG      59.60      64.35      61.42
#3  TGAAGCTGCCAGCATGATCTA     190.54     161.82     146.95
#4 CTTTGTCTATCGTTTGGAAAAG     148.04     147.29     150.32
#5  GATCATGTTCGCAGTTTCACC      82.46      73.25      88.36
#6  TTTCCAAATGTAGACAAAGCA     704.74     441.07     555.35
}
