% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sRNA_length_distribution.R
\name{srnapredata}
\alias{srnapredata}
\title{Generate the data of sRNA length distribution}
\usage{
srnapredata(sRNAseq, Group)
}
\arguments{
\item{sRNAseq}{Character. All sRNA sequences.}

\item{Group}{A character. You an select a representative group name for next drawing.}
}
\value{
A dataframe. The output results are consist of three columns, the first column is the length of sRNA, the second column id the frequency, and the third column is the group name.
}
\description{
Generally, the length interval of sRNA is 21-24. The function of "srnapredata" can provide the input data for the next drawing of sRNA length distribution among different species.
}
\examples{
#P1(B.napus)
B.napu_sRNA <- srnapredata(sRNAseq = P1_sRNA_seq,Group = "B.napus(AACC)")
#P2(B.rapa)
B.rapa_sRNA <- srnapredata(sRNAseq = P2_sRNA_seq,Group = "B.rapa(AA)")
#F1(B.napus X B.rapa)
B.nr_sRNA <- srnapredata(sRNAseq = F1_sRNA_seq,Group = "B.napus x B.rapa(AAAACC)")
#intergrate these data for length distribution plot
sRNA_data <- rbind(B.napu_sRNA,B.rapa_sRNA,B.nr_sRNA)
}
