% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpEXP.R
\name{VennData}
\alias{VennData}
\title{Species specific expression analysis: VennData}
\usage{
VennData(
  P1_RPM,
  P2_RPM,
  F1_RPM,
  type,
  homoeologs,
  rpm_threshold = 1,
  output_file = "venn_list"
)
}
\arguments{
\item{P1_RPM}{A dataframe. The rpm data of miRNA from the P1 species. The first column must be the miRNA sequence. Others are listed as the rpm of miRNA, and each column denotes one biological replicate of the sample.}

\item{P2_RPM}{A dataframe. Similar with P1_RPM, the rpm data of miRNA from the P2 species.}

\item{F1_RPM}{A dataframe. Similar with P1_RPM, the rpm data of miRNA from the F1 species.}

\item{type}{A character. "sRNA" or "mRNA".}

\item{homoeologs}{A dataframe. if the "type" is "mRNA", the option is required. The first column is the group name (unique) of homoeologs among three species, the second column is the Gene ID of P1, the third column is the Gene ID of P2. And the fourth column is the ID in the F1 generation that is consistent with the P1 genome type, while the fifth column is the ID in the F1 generation that is consistent with the P2 genome type. Such as "Homoeolog1    BraA01t00004Z    BolC01g000040.2J BnA01g0000030.1 BnC01g0424620.1".}

\item{rpm_threshold}{A numeric. the average of rpm value among all the biological replicates. By default, the average of rpm more than or equal to 1 is retained.}

\item{output_file}{Specify the output file. "venn_list" is the default option, which outputs all the information of the Venn diagram. "all_common" is one of options, which outputs the miRNAs shared by parents and offspring. "P1_specific" is one of options, which outputs P1 specific expression miRNA. "P2_specific" is one of options, which outputs P2 specific expression miRNA. "F1_specific" is one of options, which outputs F1 specific expression miRNA.}
}
\value{
A dataframe. The output results is based on your selection (output_file).
}
\description{
VennData: Extract the species-specific miRNAs and the shared miRNAs among parents and offspring.
}
\examples{
##Extract the species-specific miRNAs and the shared miRNAs among parents and offspring.
##output_file = "venn_list"
venn_list <- VennData(P1_RPM = P1_miRNA_rpm,
                        P2_RPM = P2_miRNA_rpm,
                        F1_RPM = F1_miRNA_rpm,
                        type="sRNA",rpm_threshold = 1,
                        output_file = "venn_list")
##output_file = "P1_specific"
P1_specific <- VennData(P1_RPM = P1_miRNA_rpm,
                          P2_RPM = P2_miRNA_rpm,
                          F1_RPM = F1_miRNA_rpm,
                          type="sRNA",rpm_threshold = 1,
                          output_file = "P1_specific")
##output_file = "P2_specific"
P2_specific <- VennData(P1_RPM = P1_miRNA_rpm,
                          P2_RPM = P2_miRNA_rpm,
                          F1_RPM = F1_miRNA_rpm,
                          type="sRNA",rpm_threshold = 1,
                          output_file = "P2_specific")
##output_file = "F1_specific"
F1_specific <- VennData(P1_RPM = P1_miRNA_rpm,
                          P2_RPM = P2_miRNA_rpm,
                          F1_RPM = F1_miRNA_rpm,
                          type="sRNA",rpm_threshold = 1,
                          output_file = "F1_specific")
##output_file = "all_common"
all_common <- VennData(P1_RPM = P1_miRNA_rpm,
                         P2_RPM = P2_miRNA_rpm,
                         F1_RPM = F1_miRNA_rpm,
                         type="sRNA",rpm_threshold = 1,
                         output_file = "all_common")
}
