% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpEXP.R
\name{VennPlot}
\alias{VennPlot}
\title{Species specific expression analysis: VennPlot}
\usage{
VennPlot(
  P1_RPM,
  P2_RPM,
  F1_RPM,
  P1_name,
  P2_name,
  F1_name,
  type,
  homoeologs,
  rpm_threshold = 1
)
}
\arguments{
\item{P1_RPM}{A dataframe. The rpm data of miRNA from the P1 species. The first column must be the miRNA sequence. Others are listed as the rpm of miRNA, and each column denotes one biological replicate of the sample.}

\item{P2_RPM}{A dataframe. Similar with P1_RPM, the rpm data of miRNA from the P2 species.}

\item{F1_RPM}{A dataframe. Similar with P1_RPM, the rpm data of miRNA from the F1 species.}

\item{P1_name}{A character. The text in the figure represents the name of one of the parents (P1).}

\item{P2_name}{A character. The text in the figure represents the name of one of the parents (P2).}

\item{F1_name}{A character. The text in the figure represents the name of hybrid offspring (F1).}

\item{type}{A character. "sRNA" or "mRNA".}

\item{homoeologs}{A dataframe. if the "type" is "mRNA", the option is required. The first column is the group name (unique) of homoeologs among three species, the second column is the Gene ID of P1, the third column is the Gene ID of P2. And the fourth column is the ID in the F1 generation that is consistent with the P1 genome type, while the fifth column is the ID in the F1 generation that is consistent with the P2 genome type. Such as "Homoeolog1    BraA01t00004Z    BolC01g000040.2J BnA01g0000030.1 BnC01g0424620.1".}

\item{rpm_threshold}{A numeric. the average of rpm value among all the biological replicates. By default, the average of rpm more than or equal to 1 is retained.}
}
\value{
The Venn diagram with the specific expression information of miRNAs/siRNAs/genes.
}
\description{
VennPlot: generate the Venn diagram with the specific expression information of miRNAs.
}
\examples{
##Drawing
VennPlot(P1_RPM = P1_miRNA_rpm,
           P2_RPM = P2_miRNA_rpm,
           F1_RPM = F1_miRNA_rpm,
           P1_name = "B.napus(AACC)",
           P2_name = "B.rapa(AA)",
           F1_name = "B.napus x B.rapa(AAAACC)",type="sRNA")
}
